requirements.txt
# Kafka
kafka-python==2.0.2

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# API requests
requests==2.31.0

# Dashboard
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3

# Logging/Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-mock==3.12.0

